Move ItemRange to separate file
